test(tree) move old top-level tests to appropriate area
